Fix bug in NER conversion + add tests
