+++ test/test_conversion.py
@@ -1,4 +1,9 @@
-from farm.modeling.adaptive_model import AdaptiveModel
+import pytest
+from transformers import BertForSequenceClassification, BertForTokenClassification
+
+from farm.modeling.adaptive_model import AdaptiveModel, Converter
+from farm.modeling.language_model import LanguageModel
+from farm.modeling.prediction_head import TextClassificationHead, TokenClassificationHead, QuestionAnsweringHead
 from farm.modeling.tokenization import Tokenizer
 from farm.infer import Inferencer
 from transformers.pipelines import pipeline
@@ -6,14 +11,14 @@ from transformers.modeling_auto import AutoModelForQuestionAnswering, AutoModelW
     AutoModelForSequenceClassification, AutoModelForTokenClassification
 import os
 from pathlib import Path
-from farm.conversion.converter import Converter
+
 
 import logging
 
 
 def test_conversion_adaptive_model_qa():
     farm_model = Converter.convert_from_transformers("deepset/bert-base-cased-squad2", device="cpu")
-    transformer_model = farm_model.convert_to_transformers()
+    transformer_model = farm_model.convert_to_transformers()[0]
     transformer_model2 = AutoModelForQuestionAnswering.from_pretrained("deepset/bert-base-cased-squad2")
     # compare weights
     for p1, p2 in zip(transformer_model.parameters(), transformer_model2.parameters()):
@@ -22,7 +27,7 @@ def test_conversion_adaptive_model_qa():
 
 def test_conversion_adaptive_model_lm():
     farm_model = Converter.convert_from_transformers("bert-base-german-cased", device="cpu")
-    transformer_model = farm_model.convert_to_transformers()
+    transformer_model = farm_model.convert_to_transformers()[0]
     transformer_model2 = AutoModelWithLMHead.from_pretrained("bert-base-german-cased")
     # compare weights
     for p1, p2 in zip(transformer_model.parameters(), transformer_model2.parameters()):
@@ -31,8 +36,8 @@ def test_conversion_adaptive_model_lm():
 
 def test_conversion_adaptive_model_classification():
     farm_model = Converter.convert_from_transformers("deepset/bert-base-german-cased-hatespeech-GermEval18Coarse", device="cpu")
-    transformer_model = farm_model.convert_to_transformers()
-    transformer_model2 = AutoModelForSequenceClassification.from_pretrained("ddeepset/bert-base-german-cased-hatespeech-GermEval18Coarse")
+    transformer_model = farm_model.convert_to_transformers()[0]
+    transformer_model2 = AutoModelForSequenceClassification.from_pretrained("deepset/bert-base-german-cased-hatespeech-GermEval18Coarse")
     # compare weights
     for p1, p2 in zip(transformer_model.parameters(), transformer_model2.parameters()):
         assert (p1.data.ne(p2.data).sum() == 0)
@@ -40,7 +45,7 @@ def test_conversion_adaptive_model_classification():
 
 def test_conversion_adaptive_model_ner():
     farm_model = Converter.convert_from_transformers("dslim/bert-base-NER", device="cpu")
-    transformer_model = farm_model.convert_to_transformers()
+    transformer_model = farm_model.convert_to_transformers()[0]
     transformer_model2 = AutoModelForTokenClassification.from_pretrained("dslim/bert-base-NER")
     # compare weights
     for p1, p2 in zip(transformer_model.parameters(), transformer_model2.parameters()):
@@ -147,11 +152,11 @@ def test_conversion_inferencer_ner():
     assert len(pred_farm) == 2
 
     # save it
-    farm_model_dir = "testsave/bert-conversion-test-hf"
+    farm_model_dir = Path("testsave/bert-conversion-test-hf")
     nlp.save(farm_model_dir)
 
     # load from disk in FARM format
-    model = Converter.convert_from_transformers(farm_model_dir, device="cpu")
+    model = AdaptiveModel.load(farm_model_dir, device="cpu")
     tokenizer = Tokenizer.load(farm_model_dir)
 
     # convert to transformers
@@ -166,4 +171,18 @@ def test_conversion_inferencer_ner():
     # run predictions (using transformers)
     nlp = pipeline('ner', model=model_dir, tokenizer=model_dir)
     result_transformers = nlp(text)
-    # Transformers model contains wrong labels and wrong number of labels
+    assert result_transformers[0]["entity"] == 'B-LOC'
+    assert result_transformers[1]["entity"] == 'B-LOC'
+    assert len(result_transformers) == 2
+
+
+def test_multiple_prediction_heads():
+    model = "bert-base-german-cased"
+    lm = LanguageModel.load(model)
+    ph1 = TextClassificationHead(num_labels=3, label_list=["negative", "neutral", "positive"])
+    ph2 = TokenClassificationHead(num_labels=3, label_list=["PER", "LOC", "ORG"])
+    adaptive_model = AdaptiveModel(language_model=lm, prediction_heads=[ph1, ph2], embeds_dropout_prob=0.1,
+                                   lm_output_types="per_token", device="cpu")
+    transformer_models = Converter.convert_to_transformers(adaptive_model)
+    assert isinstance(transformer_models[0], BertForSequenceClassification)
+    assert isinstance(transformer_models[1], BertForTokenClassification)
